// src/apps/gestor-fichas/lib/types.ts
export type Ambito = "UE" | "ESTADO" | "CCAA" | "PROVINCIA" | "";
export type TramiteTipo = "si" | "no" | "directo" | "";
export type Complejidad = "baja" | "media" | "alta" | "";
// ...
export type Filters = {
  q?: string;

  ambito?: Ambito;
  ccaa_id?: string;
  provincia_id?: string;
  provincia_principal?: string;  // Filtro principal de provincia (inclusivo)

  tramite_tipo?: TramiteTipo;
  complejidad?: Complejidad;
  
  tematica_id?: string;

  trabajador_id?: string;
  trabajador_subida_id?: string;

  anio?: string;        // "2025"
  mes?: string;         // "1".."12" o ""

  created_desde?: string;  // "YYYY-MM-DD"
  created_hasta?: string;  // "YYYY-MM-DD"

  take?: string;
  page?: string;

  /** opcional si lo usas en filtros avanzados */
  existe_frase?: "" | "true" | "false";
  
  /** filtros por etiquetas de destaque */
  destaque_principal?: "" | "nueva" | "para_publicitar" | "ambas" | "no_nueva" | "no_para_publicitar";
  destaque_secundario?: "" | "para_publicitar" | "no_para_publicitar";
};